Revert "wip: adds filter viewer JavaFX utility application for visualizing filters designed through the filter factory."
This reverts commit 097884337bab3044061783bad42d07520acfdfcb.

+++ src/dsp/filter/GoertzelFilter.java
@@ -32,7 +32,7 @@ public class GoertzelFilter
     private int mSampleRate;
     private long mTargetFrequency;
     private int mBlockSize;
-    private float[] mWindowCoefficients;
+    private double[] mWindowCoefficients;
 
     private double mCoefficient;
